requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
pillow==10.2.0
folium==0.15.1
streamlit-folium==0.18.0
sqlite3 